Create POE_PART_ONE.yml
The changes that i made are for the two sub classes which are ASCIIART, where i added the color into my picture and i used three colors for the picture to be a bit nice. Then the Chatbot class, i changed the color for the chatbot replies and the user replies.